Logistics: shared shipment-detail modal, order-suggestions/inbound-log fixes, and a systemic Dapper int/bigint audit
- Extract the Inbound Shipment detail modal out of inbound-log.js into a
  new shared wwwroot/js/logistics/shipment-detail.js (window.ShipmentDetailModal.open),
  so Order Suggestions' Tracked Orders can open the same modal from its
  shipment reference link. A Received (non-cancelled) shipment now renders
  read-only with only an Undo Received action available, matching the
  Node original.
- Fix a second, different "Canvas is already in use" race on Stock History
  & Forecast's single-material chart (a generation counter plus
  Chart.getChart() backstops), on top of the earlier vendor-grid fix.
- Systemic SQL audit: dumped the live schema and found several real
  Dapper strict-materialization bugs where a record's ID field was
  declared `long`/`long?` against a genuinely `int` column
  (log.PurchaseOrderShipment.ShipmentId, log.PurchaseOrderSuggestion.
  SuggestionId, log.Vendor.VendorId, log.VendorMaterial.VendorMaterialId,
  log.DemandAdjustment.AdjustmentId, log.IsoparMeterReading.ReadingId,
  log.IsoparPlanningRate.RateId, log.IsoparDeclaration.DeclarationId, and
  every Vendor Consignment Tracker ID). These throw regardless of row
  content, so every Create Vendor/Vendor Material/Demand Adjustment/
  Isopar reading or rate/Isopar declaration, Accept/Delete order
  suggestion, Create (manual) Inbound Shipment, and Add Associated Cost
  request was broken. All fixed; full 699-test suite still green.
- Earlier this session: MRP Analysis fixes, the Stock History & Forecast
  tile (single-material + vendor-grid views, MRP/Demand Adjustment quick
  links), demand adjustment fixed-quantity override mode, Tracked Orders
  readability + Assign Shipment/completed-and-assigned lockdown, and the
  Build Order / Inbound Log features.

+++ dotnet/NormantonNexus.Tests/Helpers/Logistics/ConsignmentTrackerHelperTests.cs
@@ -10,7 +10,7 @@ namespace NormantonNexus.Tests.Helpers.Logistics;
 // documented production incidents (Raaj Ratna reversal chains).
 public class ConsignmentTrackerHelperTests
 {
-    private static ReversalWalkRow Row(long id, string doc, string item, decimal qty, decimal remaining, string? reversalDoc = null, string? reversalItem = null) =>
+    private static ReversalWalkRow Row(int id, string doc, string item, decimal qty, decimal remaining, string? reversalDoc = null, string? reversalItem = null) =>
         new(id, "MAT001", doc, item, qty, remaining, reversalDoc, reversalItem);
 
     [Fact]
@@ -108,7 +108,7 @@ public class ConsignmentTrackerHelperTests
         Assert.Empty(result.NeedsReview);
     }
 
-    private static AllocatableDeliveryRow Allocatable(long id, decimal remaining, DateTime? expiry = null, DateTime? documentDate = null) =>
+    private static AllocatableDeliveryRow Allocatable(int id, decimal remaining, DateTime? expiry = null, DateTime? documentDate = null) =>
         new(id, "MAT001", remaining, $"INV{id}", expiry, documentDate);
 
     [Fact]
@@ -169,7 +169,7 @@ public class ConsignmentTrackerHelperTests
         Assert.Equal(5.679m, result.Lines[0].QtyAllocated);
     }
 
-    private static OpenDeliveryForReassignment OpenDelivery(long id, decimal remaining, DateTime? expiry = null, DateTime? documentDate = null) =>
+    private static OpenDeliveryForReassignment OpenDelivery(int id, decimal remaining, DateTime? expiry = null, DateTime? documentDate = null) =>
         new(id, "MAT001", remaining, expiry, documentDate);
 
     [Fact]
